+++ src/assets/styles/index.ts
@@ -37,3 +37,20 @@ export const text = css`
 export const boldText = css`
   font-family: 'BCSans-Bold';
 `;
+
+export const paginationStyle = css`
+  position: relative;
+`;
+
+export const paginationStyleItemActive = css`
+  width: 10px;
+  height: 10px;
+  background-color: ${theme.colors.primaryBlue};
+`;
+
+export const paginationStyleItemInactive = css`
+  width: 10px;
+  height: 10px;
+  background-color: ${theme.colors.transparent};
+  border: 1px solid ${theme.colors.primaryBlue};
+`;
